feat(v033): RFC 0024 PR 5 re-review follow-up — settle in-flight handle on hard-cancel
- (r4) In-flight FIFO handle settled on hard-cancel: `_drain_pending_handles`
  only reaches *queued* wakes, so a handle-bearing wake whose `on_event`
  outlived the stop timeout was orphaned when `stop()` hard-cancelled the
  supervisor mid-wake (the awaiting caller then hung to its own external
  `wait_for` deadline). `_handle_wake_supervised`'s CancelledError branch now
  resolves a handle-bearing wake with the empty discard result before
  re-raising — the in-flight counterpart to the drain, parity with the
  reentrant resolver's cancellation path. Idempotent, confined to the
  already-degraded stop-timeout path. Pinned by
  `test_stop_settles_in_flight_handle_on_hard_cancel`.
- (r5) Documented `stop()`'s worst-case ~2x timeout (two sequential
  bounded phases).
- Restored `_evt`/`_build_loop` type annotations in
  test_event_loop_lifecycle.py to match the sibling green-path file.
- Split the reentrant-resolver observability-parity tests into
  test_event_loop_observability.py so the lifecycle file stays under the
  500-line pre-commit gate after the new test landed.
- PR 5.1 checklist: added `event_loop_salience._decide` to the lazy-import
  hoist cleanup; recorded (r4)/(r5) dispositions in the plan.
