+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-05-22  
-> **1027** tracked files
+> **1028** tracked files
 
 ## Summary
 
@@ -10,7 +10,7 @@
 | `internal/` | 239 |
 | `docs/` | 229 |
 | `tests/` | 228 |
-| `agents/` | 184 |
+| `agents/` | 185 |
 | `prompts/` | 28 |
 | `(root)` | 23 |
 | `scripts/` | 23 |
@@ -27,7 +27,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 426 |
+| `.py` | 427 |
 | `.md` | 272 |
 | `.go` | 250 |
 | `.yaml` | 27 |
@@ -244,6 +244,7 @@
 │   │   ├── test_event_loop_inbound.py
 │   │   ├── test_event_loop_initial_delay.py
 │   │   ├── test_event_loop_lifecycle.py
+│   │   ├── test_event_loop_observability.py
 │   │   ├── test_event_loop_salience.py
 │   │   ├── test_event_loop_salience_default_off.py
 │   │   ├── test_event_loop_timers.py
