+++ agents/tests/test_event_loop_lifecycle.py
@@ -15,10 +15,11 @@ Pins the contracts added in PR 5:
   the per-drop WARNING is emitted at most once per
   ``_QUEUE_FULL_LOG_INTERVAL_S``.
 
-* **Reentrant resolver observability parity**
-  (:class:`TestReentrantDispatchObservability`) — the detached resolver
-  records ``agent.wake.inbound`` and logs ``wake_failed`` on a raising
-  ``on_event``, matching the supervised FIFO path (PR 5 re-review).
+Reentrant-resolver observability-parity tests (the ``agent.wake.inbound``
+counter + ``wake_failed`` log) live in the sibling
+``test_event_loop_observability.py`` — split off so this file (pure
+correctness contracts) stays under the 500-line limit once the
+in-flight-handle hard-cancel test landed.
 
 Split out of ``test_event_loop.py`` for file-size review-friendliness
 (RFC 0024 PR 5 — both files exceeded the 500-line limit).
@@ -28,11 +29,8 @@ from __future__ import annotations
 
 import asyncio
 import logging
-from collections.abc import Iterator
+from collections.abc import Awaitable, Callable
 from typing import Any
-
-import pytest
-from opentelemetry.sdk.metrics.export import InMemoryMetricReader
 
 from agents.event_loop import (
     EventLoop,
@@ -40,7 +38,6 @@ from agents.event_loop import (
     ScheduledWake,
     SyncDispatchHandle,
 )
-from agents.observability import metrics as pmetrics
 from agents.persona_types import (
     ActionType,
     AgentAction,
@@ -49,36 +46,7 @@ from agents.persona_types import (
 )
 
 
-@pytest.fixture
-def metric_reader() -> Iterator[InMemoryMetricReader]:
-    """An in-memory OTEL reader wired into the process-global instruments.
-
-    Mirrors the fixture in ``test_observability_metrics.py`` — the
-    EventLoop records the ``agent.wake.*`` counters via the module-global
-    instruments, so a test that asserts on them must initialise the
-    metrics provider and tear it down afterwards.
-    """
-    reader = InMemoryMetricReader()
-    pmetrics.init_metrics(reader=reader)
-    try:
-        yield reader
-    finally:
-        asyncio.run(pmetrics.shutdown())
-
-
-def _collect(reader: InMemoryMetricReader) -> dict[str, Any]:
-    data = reader.get_metrics_data()
-    out: dict[str, Any] = {}
-    if data is None:
-        return out
-    for rm in data.resource_metrics:
-        for sm in rm.scope_metrics:
-            for m in sm.metrics:
-                out[m.name] = m
-    return out
-
-
-def _evt(payload: dict | None = None) -> AgentEvent:
+def _evt(payload: dict[str, Any] | None = None) -> AgentEvent:
     return AgentEvent(
         event_type=EventType.CHANNEL_MESSAGE,
         payload=payload or {"content": "hi"},
@@ -87,8 +55,8 @@ def _evt(payload: dict | None = None) -> AgentEvent:
 
 def _build_loop(
     *,
-    on_event=None,
-    on_tick=None,
+    on_event: Callable[[AgentEvent], Awaitable[list[AgentAction]]] | None = None,
+    on_tick: Callable[[ScheduledWake], Awaitable[None]] | None = None,
     queue_size: int = 1024,
     agent_id: str = "test-agent",
 ) -> EventLoop:
@@ -360,113 +328,44 @@ class TestStopSettlesPendingHandles:
             if not block.is_set():
                 block.set()
 
+    async def test_stop_settles_in_flight_handle_on_hard_cancel(self):
+        """A handle-bearing FIFO wake whose ``on_event`` is still running when
+        ``stop()`` hard-cancels the supervisor (stop timeout exceeded) has its
+        handle settled with the empty discard result, not orphaned.
 
-class TestReentrantDispatchObservability:
-    """RFC 0024 PR 5 re-review (review finding #1) — the reentrant resolver
-    must have observability parity with the supervised FIFO path.
-
-    The FIFO path records the ``agent.wake.inbound`` counter for every
-    inbound wake and logs ``agent.event_loop.wake_failed`` (with a
-    traceback) when ``on_event`` raises.  The reentrant escape hatch runs
-    ``on_event`` on a detached resolver task that bypasses both, so a
-    reentrant inbound wake was invisible to the inbound counter and a
-    reentrant failure was rejected silently.  These tests pin the parity.
-    """
-
-    async def test_reentrant_inbound_wake_records_inbound_metric(
-        self, metric_reader: InMemoryMetricReader,
-    ):
-        """A reentrant inbound wake increments ``agent.wake.inbound`` just
-        like a FIFO-drained one.  The reentrant wake is triggered from a
-        ``ScheduledWake`` (which records ``agent.wake.scheduled``, not
-        inbound), so the *only* contributor to this agent's inbound counter
-        is the reentrant resolver — isolating the behaviour under test.
+        Unlike the queued wakes :meth:`_drain_pending_handles` settles, the
+        *in-flight* wake has already been dequeued, so the drain cannot reach
+        it.  When ``on_event`` outlives the stop timeout the supervisor is
+        cancelled mid-wake; the cancellation must settle the handle with
+        ``[]`` — parity with the reentrant resolver's cancellation path
+        (:meth:`_spawn_reentrant`) and the queued-wake drain — so the awaiting
+        caller resolves immediately instead of hanging to its own external
+        ``wait_for`` deadline. (PR 5 re-review.)
         """
-        done = asyncio.Event()
-        loop_ref: list[EventLoop] = []
-
-        async def _on_tick(wake: ScheduledWake) -> None:
-            # Reentrant handle-bearing inbound enqueue from inside on_tick.
-            handle = SyncDispatchHandle()
-            loop_ref[0].enqueue(
-                InboundEventWake(event=_evt({"level": "inner"}), handle=handle),
-            )
-            await handle
-            done.set()
-
-        loop = _build_loop(on_tick=_on_tick, agent_id="reentrant-metric")
-        loop_ref.append(loop)
-        loop.start()
-        try:
-            loop.enqueue(ScheduledWake(timer_id="t", callback_kind="tick"))
-            await asyncio.wait_for(done.wait(), timeout=2.0)
-        finally:
-            await loop.stop(timeout=1.0)
-
-        m = _collect(metric_reader).get("agent.wake.inbound")
-        assert m is not None, (
-            "reentrant inbound wake must record agent.wake.inbound"
-        )
-        total = sum(
-            getattr(dp, "value", 0)
-            for dp in m.data.data_points
-            if dict(dp.attributes).get("agent.id") == "reentrant-metric"
-        )
-        assert total == 1
-
-    async def test_reentrant_on_event_failure_logs_and_rejects(self, caplog):
-        """A reentrant resolver whose ``on_event`` raises rejects the handle
-        *and* logs ``wake_failed`` with a traceback — parity with the
-        supervised FIFO path.  The outer body swallows the rejection so the
-        only ``wake_failed`` record can come from the reentrant resolver.
-        """
-        boom = RuntimeError("reentrant boom")
-        inner_handle_ref: list[SyncDispatchHandle] = []
-        loop_ref: list[EventLoop] = []
+        block = asyncio.Event()
+        in_flight = asyncio.Event()
 
         async def _on_event(event: AgentEvent) -> list[AgentAction]:
-            if event.payload.get("level") == "outer":
-                inner_handle = SyncDispatchHandle()
-                inner_handle_ref.append(inner_handle)
-                loop_ref[0].enqueue(
-                    InboundEventWake(
-                        event=_evt({"level": "inner"}), handle=inner_handle,
-                    ),
-                )
-                try:
-                    return await inner_handle
-                except RuntimeError:
-                    # Swallow so the supervised path does not also log a
-                    # wake_failed for the outer wake — isolates the resolver.
-                    return [AgentAction(ActionType.DO_NOTHING, {})]
-            raise boom
+            in_flight.set()
+            # Never released: forces stop() to hit its timeout and
+            # hard-cancel the supervisor while this wake is in flight.
+            await block.wait()
+            return [AgentAction(ActionType.DO_NOTHING, {})]
 
         loop = _build_loop(on_event=_on_event)
-        loop_ref.append(loop)
         loop.start()
         try:
-            outer_handle = SyncDispatchHandle()
-            with caplog.at_level(logging.ERROR, logger="agents.event_loop"):
-                loop.enqueue(
-                    InboundEventWake(
-                        event=_evt({"level": "outer"}), handle=outer_handle,
-                    ),
-                )
-                # Outer resolves with the swallowed result once the inner
-                # rejection has propagated through the resolver.
-                assert await asyncio.wait_for(outer_handle, timeout=0.5) == [
-                    AgentAction(ActionType.DO_NOTHING, {}),
-                ]
-            # Resolver rejected the inner handle...
-            assert inner_handle_ref[0].done()
-            with pytest.raises(RuntimeError, match="reentrant boom"):
-                await inner_handle_ref[0]
-            # ...and logged the failure with a traceback (parity with the
-            # supervised FIFO path). Without the fix the rejection is silent.
-            failed = [
-                r for r in caplog.records if "wake_failed" in r.getMessage()
-            ]
-            assert len(failed) == 1
-            assert failed[0].exc_info is not None
+            handle = SyncDispatchHandle()
+            assert loop.enqueue(InboundEventWake(event=_evt(), handle=handle))
+            await asyncio.wait_for(in_flight.wait(), timeout=2.0)
+
+            # Short timeout: on_event blocks past it, so stop() hard-cancels
+            # the supervisor.  The in-flight handle must be settled by the
+            # time stop() returns.
+            await loop.stop(timeout=0.1)
+
+            assert handle.done()
+            assert await asyncio.wait_for(handle, timeout=0.5) == []
         finally:
-            await loop.stop(timeout=1.0)
+            if not block.is_set():
+                block.set()
