+++ docs/rfcs/0024-pr-plan.md
@@ -404,6 +404,8 @@ _**From the PR 5 re-review** (follow-ups on the lifecycle-hardening diff itself 
 - _**(r1) Reentrant resolver observability parity.** The detached `_spawn_reentrant` resolver bypassed `_handle_wake`, so a reentrant `InboundEventWake` was invisible to the `agent.wake.inbound` counter and a raising reentrant `on_event` was rejected silently (the supervised FIFO path logs `agent.event_loop.wake_failed` with a traceback). `_spawn_reentrant` now records `agent.wake.inbound` and logs `wake_failed` with `exc_info` before rejecting. Pinned by `TestReentrantDispatchObservability` (metric + log+reject)._
 - _**(r2) `stop()` ordering — reentrant resolvers cancelled and awaited first.** The original `stop()` cancelled in-flight reentrant resolvers *after* the supervisor wait and never awaited them, so (a) a supervisor parked on a reentrant handle waited out the full `timeout` and was hard-cancelled — which cascaded into the awaited future and left the handle *cancelled* rather than resolved — and (b) the handle settled only on a later tick, not before `stop()` returns. The cancel + `await asyncio.gather(..., return_exceptions=True)` now runs before the supervisor wait; the already-set `_stopped` guard prevents any new reentrant spawn, so the snapshot is complete. Pinned by `test_stop_settles_in_flight_reentrant_handle`._
 - _**(r3) Single-level-reentrancy invariant documented.** `_spawn_reentrant` runs `on_event` (decide-only), which does not itself re-dispatch; a reentrant enqueue from *inside* the detached resolver would fail the `current_task() is self._task` check and land on the FIFO the blocked supervisor cannot drain. The escape hatch is correct only while the detached body stays decide-only — now stated in the `_spawn_reentrant` docstring so a future change cannot silently reintroduce the deadlock. `resolved (docstring only)`._
+- _**(r4) In-flight FIFO handle settled on hard-cancel.** `_drain_pending_handles` settles only *queued* handles; a handle-bearing wake whose `on_event` outlived the stop timeout was already dequeued, so when `stop()` hard-cancelled the supervisor mid-wake its handle was orphaned (the awaiting caller then hung to its own external `wait_for` deadline rather than the graceful `[]` discard). `_handle_wake_supervised`'s `CancelledError` branch now resolves a handle-bearing wake with `[]` before re-raising — the in-flight counterpart to the drain and parity with the reentrant resolver's cancellation path. Confined to the already-degraded stop-timeout path; `resolve` is idempotent so a normally-completed wake is unaffected. Pinned by `test_stop_settles_in_flight_handle_on_hard_cancel`._
+- _**(r5) `stop()` worst-case bound documented.** `timeout` bounds the reentrant-cancel wait and the supervisor-drain wait independently and sequentially, so worst-case wall time is up to ~2× `timeout` — but only if both phases hit their bound (a handler swallowing `CancelledError`). Now stated in the `stop()` docstring. `resolved (docstring only)`._
 
 ##### From PR 2 review
 
@@ -685,7 +687,7 @@ _Deferred:_
 
 **PR 5.1 (cleanups + test gaps):**
 - [ ] Remaining *needed* findings addressed; conditional-defer findings confirmed downgraded with rationale.
-- [ ] Hoist the runtime `from .event_loop_types import …` imports inside `event_loop_lifecycle._drain_pending_handles` and `event_loop_timers._arm_timer` to module scope. The "local: avoid import cycle" comments predate the PR 5 split: now that the wake taxonomy lives in `event_loop_types` (which imports neither sibling), the cycle is gone and the lazy imports are dead defensiveness (PR 5 re-review nit). `→ PR 5.1`.
+- [ ] Hoist the runtime wake-taxonomy imports to module scope in the three remaining lazy-import sites: `event_loop_lifecycle._drain_pending_handles` and `event_loop_timers._arm_timer` (both already `from .event_loop_types import …`), plus `event_loop_salience._decide` (still `from .event_loop import SalienceWake` — re-point to `event_loop_types` then hoist). The "local: avoid import cycle" comments predate the PR 5 split: now that the wake taxonomy lives in `event_loop_types` (which imports neither sibling), the cycle is gone and the lazy imports are dead defensiveness (PR 5 re-review nit). `→ PR 5.1`.
 - [ ] `make test` + `make lint` clean.
 - [ ] Deferred test gaps from PRs 1–4 reviews filled (PR 2 (8) precedence regression; PR 2.1 (3) banker's-rounding).
 - [ ] [Progress Overview](#progress-overview) row 5.1 filled.
